Split synapse.rs (4,375 lines) into focused submodules (#482)
Pure structural refactoring — no logic changes. Splits the monolithic
synapse analysis module into focused submodules under src/analysis/synapse/:

- mod.rs: Core pipeline, public API, re-exports, tests (2,485 lines)
- gpu_evaluation.rs: ReLU/activation evaluation, batched GPU (975 lines)
- scoring.rs: Improvement calculation, boosting (532 lines)
- filtering.rs: Candidate truncation, deduplication (243 lines)
- candidate_generation.rs: Sample locality, ordered neurons (220 lines)

All 463+ existing tests pass. No public API changes.
